+++ solr/src/main/scala/akka/stream/alpakka/solr/impl/SolrFlowStage.scala
@@ -1,0 +1,206 @@
+package akka.stream.alpakka.solr.impl
+
+import akka.annotation.InternalApi
+import akka.stream.ActorAttributes.SupervisionStrategy
+import akka.stream._
+import akka.stream.alpakka.solr._
+import akka.stream.stage._
+import org.apache.solr.client.solrj.SolrClient
+import org.apache.solr.client.solrj.impl.CloudSolrClient
+import org.apache.solr.client.solrj.response.UpdateResponse
+import org.apache.solr.common.SolrInputDocument
+
+import scala.annotation.tailrec
+import scala.util.control.NonFatal
+
+/**
+ * Internal API
+ *
+ * @tparam T
+ * @tparam C pass-through type
+ */
+@InternalApi
+private[solr] final class SolrFlowStage[T, C](
+    collection: String,
+    client: SolrClient,
+    settings: SolrUpdateSettings,
+    messageBinder: T => SolrInputDocument
+) extends GraphStage[FlowShape[Seq[IncomingMessage[T, C]], Seq[IncomingMessageResult[T, C]]]] {
+
+  private val in = Inlet[Seq[IncomingMessage[T, C]]]("messages")
+  private val out = Outlet[Seq[IncomingMessageResult[T, C]]]("result")
+  override val shape = FlowShape(in, out)
+
+  override protected def initialAttributes: Attributes =
+    super.initialAttributes and Attributes(ActorAttributes.IODispatcher)
+
+  override def createLogic(inheritedAttributes: Attributes): SolrFlowLogic[T, C] = {
+    def decider = inheritedAttributes.mandatoryAttribute[SupervisionStrategy].decider
+    new SolrFlowLogic[T, C](decider, collection, client, in, out, shape, settings, messageBinder)
+  }
+}
+
+private final class SolrFlowLogic[T, C](
+    decider: Supervision.Decider,
+    collection: String,
+    client: SolrClient,
+    in: Inlet[Seq[IncomingMessage[T, C]]],
+    out: Outlet[Seq[IncomingMessageResult[T, C]]],
+    shape: FlowShape[Seq[IncomingMessage[T, C]], Seq[IncomingMessageResult[T, C]]],
+    settings: SolrUpdateSettings,
+    messageBinder: T => SolrInputDocument
+) extends GraphStageLogic(shape)
+    with OutHandler
+    with InHandler
+    with StageLogging {
+
+  setHandlers(in, out, this)
+
+  override def onPull(): Unit =
+    tryPull()
+
+  override def onPush(): Unit = {
+    val messagesIn = grab(in)
+    try {
+      sendBulkToSolr(messagesIn)
+      tryPull()
+    } catch {
+      case NonFatal(ex) ⇒
+        decider(ex) match {
+          case Supervision.Stop ⇒ failStage(ex)
+          case _ ⇒ tryPull() // for resume and restart strategies tryPull
+        }
+    }
+
+  }
+
+  override def preStart(): Unit =
+    pull(in)
+
+  override def onUpstreamFailure(ex: Throwable): Unit =
+    handleFailure(ex)
+
+  override def onUpstreamFinish(): Unit = handleSuccess
+
+  private def tryPull(): Unit =
+    if (!isClosed(in) && !hasBeenPulled(in)) {
+      pull(in)
+    }
+
+  private def handleFailure(exc: Throwable): Unit = {
+    log.warning(s"Received error from solr. Error: ${exc.toString}")
+    failStage(exc)
+  }
+
+  private def handleResponse(args: (Seq[IncomingMessage[T, C]], Int)): Unit = {
+    val (messages, status) = args
+    log.debug(s"Handle the response with $status")
+    val result = messages.map(
+      m =>
+        IncomingMessageResult(m.idField,
+                              m.idFieldValue,
+                              m.routingFieldValue,
+                              m.query,
+                              m.source,
+                              m.updates,
+                              m.passThrough,
+                              status)
+    )
+
+    emit(out, result)
+  }
+
+  private def handleSuccess(): Unit =
+    completeStage()
+
+  private def updateBulkToSolr(messages: Seq[IncomingMessage[T, C]]): UpdateResponse = {
+    val docs = messages.flatMap(_.source.map(messageBinder))
+
+    if (log.isDebugEnabled) log.debug(s"Upsert $docs")
+    client.add(collection, docs.asJava, settings.commitWithin)
+  }
+
+  private def atomicUpdateBulkToSolr(messages: Seq[IncomingMessage[T, C]]): UpdateResponse = {
+    val docs = messages.map { message =>
+      val doc = new SolrInputDocument()
+
+      message.idField.foreach { idField =>
+        message.idFieldValue.foreach { idFieldValue =>
+          doc.addField(idField, idFieldValue)
+        }
+      }
+
+      message.routingFieldValue.foreach { routingFieldValue =>
+        val routingField = client match {
+          case csc: CloudSolrClient =>
+            Option(csc.getIdField)
+          case _ => None
+        }
+        routingField.foreach { routingField =>
+          message.idField.foreach { idField =>
+            if (routingField != idField)
+              doc.addField(routingField, routingFieldValue)
+          }
+        }
+      }
+
+      message.updates.foreach {
+        case (field, updates) => {
+          val jMap = updates.asInstanceOf[Map[String, Any]].asJava
+          doc.addField(field, jMap)
+        }
+      }
+      doc
+    }
+    if (log.isDebugEnabled) log.debug(s"Update atomically $docs")
+    client.add(collection, docs.asJava, settings.commitWithin)
+  }
+
+  private def deleteBulkToSolrByIds(messages: Seq[IncomingMessage[T, C]]): UpdateResponse = {
+    val docsIds = messages
+      .filter { message =>
+        message.operation == DeleteByIds && message.idFieldValue.isDefined
+      }
+      .map { message =>
+        message.idFieldValue.get
+      }
+    if (log.isDebugEnabled) log.debug(s"Delete the ids $docsIds")
+    client.deleteById(collection, docsIds.asJava, settings.commitWithin)
+  }
+
+  private def deleteEachByQuery(messages: Seq[IncomingMessage[T, C]]): UpdateResponse = {
+    val responses = messages.map { message =>
+      val query = message.query.get
+      if (log.isDebugEnabled) log.debug(s"Delete by the query $query")
+      client.deleteByQuery(collection, query, settings.commitWithin)
+    }
+    responses.find(_.getStatus != 0).getOrElse(responses.head)
+  }
+
+  private def sendBulkToSolr(messages: Seq[IncomingMessage[T, C]]): Unit = {
+
+    @tailrec
+    def send(toSend: Seq[IncomingMessage[T, C]]): UpdateResponse = {
+      val operation = toSend.head.operation
+      //Just take a subset of this operation
+      val (current, remaining) = toSend.span { m =>
+        m.operation == operation
+      }
+      //send this subset
+      val response = operation match {
+        case Upsert => updateBulkToSolr(current)
+        case AtomicUpdate => atomicUpdateBulkToSolr(current)
+        case DeleteByIds => deleteBulkToSolrByIds(current)
+        case DeleteByQuery => deleteEachByQuery(current)
+      }
+      if (remaining.nonEmpty) {
+        send(remaining)
+      } else {
+        response
+      }
+    }
+
+    val response = if (messages.nonEmpty) send(messages) else new UpdateResponse
+    handleResponse((messages, response.getStatus))
+  }
+}
